UserController update
UserController allowed to do actions on users after expired session

+++ src/Controllers/UserController.php
@@ -31,21 +31,29 @@ class UserController extends Controller
 
     public function enable(int $userId)
     {
+        if ($this->authentication === false){return;}
+        
         $this->model->enable($userId);
     }
 
     public function disable(int $userId)
     {
+        if ($this->authentication === false){return;}
+        
         $this->model->disable($userId);
     }
 
     public function delete(int $userId)
     {
+        if ($this->authentication === false){return;}
+        
         $this->model->delete($userId);
     }
 
     public function changePassword(int $userId)
     {
+        if ($this->authentication === false){return;}
+        
         $newPassword = filter_input(INPUT_POST, 'new_password', FILTER_SANITIZE_STRING);
         $confirmPassword = filter_input(INPUT_POST, 'confirm_password', FILTER_SANITIZE_STRING);
         
@@ -54,6 +62,8 @@ class UserController extends Controller
 
     public function modify(int $userId)
     {
+        if ($this->authentication === false){return;}
+        
         $newUserName = filter_input(INPUT_POST, 'new_user_name', FILTER_SANITIZE_STRING);
         $newUserDescription = filter_input(INPUT_POST, 'new_user_description', FILTER_SANITIZE_STRING);
